+++ Front-End/.eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\Users\\magalu\\Desktop\\ConnectPets\\Front-End\\src\\index.js":"1","C:\\Users\\magalu\\Desktop\\ConnectPets\\Front-End\\src\\components\\Form\\styles.js":"2","C:\\Users\\magalu\\Desktop\\ConnectPets\\Front-End\\src\\components\\Pets\\styles.js":"3","C:\\Users\\magalu\\Desktop\\ConnectPets\\Front-End\\src\\components\\DropZone\\styles.js":"4","C:\\Users\\magalu\\Desktop\\ConnectPets\\Front-End\\src\\components\\DropZone\\index.js":"5","C:\\Users\\magalu\\Desktop\\ConnectPets\\Front-End\\src\\components\\Form\\index.js":"6","C:\\Users\\magalu\\Desktop\\ConnectPets\\Front-End\\src\\styles\\GlobalStyles.js":"7","C:\\Users\\magalu\\Desktop\\ConnectPets\\Front-End\\src\\components\\NavBar\\index.js":"8","C:\\Users\\magalu\\Desktop\\ConnectPets\\Front-End\\src\\components\\NavBar\\styles.js":"9","C:\\Users\\magalu\\Desktop\\ConnectPets\\Front-End\\src\\components\\Pets\\index.js":"10","C:\\Users\\magalu\\Desktop\\ConnectPets\\Front-End\\src\\routes.js":"11","C:\\Users\\magalu\\Desktop\\ConnectPets\\Front-End\\src\\services\\api.js":"12","C:\\Users\\magalu\\Desktop\\ConnectPets\\Front-End\\src\\Pages\\Home\\index.js":"13","C:\\Users\\magalu\\Desktop\\ConnectPets\\Front-End\\src\\components\\Feed\\index.js":"14","C:\\Users\\magalu\\Desktop\\ConnectPets\\Front-End\\src\\components\\Feed\\styles.js":"15","C:\\Users\\magalu\\Desktop\\ConnectPets\\Front-End\\src\\Pages\\Donate\\index.js":"16","C:\\Users\\magalu\\Desktop\\ConnectPets\\Front-End\\src\\App.js":"17"},{"size":207,"mtime":1616620331159,"results":"18","hashOfConfig":"19"},{"size":1471,"mtime":1616805748569,"results":"20","hashOfConfig":"19"},{"size":948,"mtime":1617245086666,"results":"21","hashOfConfig":"19"},{"size":617,"mtime":1616620331121,"results":"22","hashOfConfig":"19"},{"size":1032,"mtime":1617052183292,"results":"23","hashOfConfig":"19"},{"size":3328,"mtime":1617197563456,"results":"24","hashOfConfig":"19"},{"size":1053,"mtime":1616620331201,"results":"25","hashOfConfig":"19"},{"size":788,"mtime":1616620331143,"results":"26","hashOfConfig":"19"},{"size":1393,"mtime":1616620331145,"results":"27","hashOfConfig":"19"},{"size":1148,"mtime":1617245368014,"results":"28","hashOfConfig":"19"},{"size":370,"mtime":1616620331179,"results":"29","hashOfConfig":"19"},{"size":121,"mtime":1616620331193,"results":"30","hashOfConfig":"19"},{"size":175,"mtime":1616620331087,"results":"31","hashOfConfig":"19"},{"size":333,"mtime":1616620331126,"results":"32","hashOfConfig":"19"},{"size":307,"mtime":1616620331129,"results":"33","hashOfConfig":"19"},{"size":175,"mtime":1616620331080,"results":"34","hashOfConfig":"19"},{"size":360,"mtime":1616620331074,"results":"35","hashOfConfig":"19"},{"filePath":"36","messages":"37","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"38"},"1037t6n",{"filePath":"39","messages":"40","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"41","messages":"42","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"43","messages":"44","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"45","messages":"46","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"47","messages":"48","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"49","messages":"50","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"51"},{"filePath":"52","messages":"53","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"54","messages":"55","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"51"},{"filePath":"56","messages":"57","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"58","messages":"59","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"51"},{"filePath":"60","messages":"61","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"62","messages":"63","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"51"},{"filePath":"64","messages":"65","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"66","messages":"67","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"68","messages":"69","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"70","messages":"71","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\magalu\\Desktop\\ConnectPets\\Front-End\\src\\index.js",[],["72","73"],"C:\\Users\\magalu\\Desktop\\ConnectPets\\Front-End\\src\\components\\Form\\styles.js",[],"C:\\Users\\magalu\\Desktop\\ConnectPets\\Front-End\\src\\components\\Pets\\styles.js",[],"C:\\Users\\magalu\\Desktop\\ConnectPets\\Front-End\\src\\components\\DropZone\\styles.js",[],"C:\\Users\\magalu\\Desktop\\ConnectPets\\Front-End\\src\\components\\DropZone\\index.js",[],"C:\\Users\\magalu\\Desktop\\ConnectPets\\Front-End\\src\\components\\Form\\index.js",[],"C:\\Users\\magalu\\Desktop\\ConnectPets\\Front-End\\src\\styles\\GlobalStyles.js",[],["74","75"],"C:\\Users\\magalu\\Desktop\\ConnectPets\\Front-End\\src\\components\\NavBar\\index.js",[],"C:\\Users\\magalu\\Desktop\\ConnectPets\\Front-End\\src\\components\\NavBar\\styles.js",[],"C:\\Users\\magalu\\Desktop\\ConnectPets\\Front-End\\src\\components\\Pets\\index.js",[],"C:\\Users\\magalu\\Desktop\\ConnectPets\\Front-End\\src\\routes.js",[],"C:\\Users\\magalu\\Desktop\\ConnectPets\\Front-End\\src\\services\\api.js",[],"C:\\Users\\magalu\\Desktop\\ConnectPets\\Front-End\\src\\Pages\\Home\\index.js",[],"C:\\Users\\magalu\\Desktop\\ConnectPets\\Front-End\\src\\components\\Feed\\index.js",[],"C:\\Users\\magalu\\Desktop\\ConnectPets\\Front-End\\src\\components\\Feed\\styles.js",[],"C:\\Users\\magalu\\Desktop\\ConnectPets\\Front-End\\src\\Pages\\Donate\\index.js",[],"C:\\Users\\magalu\\Desktop\\ConnectPets\\Front-End\\src\\App.js",[],{"ruleId":"76","replacedBy":"77"},{"ruleId":"78","replacedBy":"79"},{"ruleId":"76","replacedBy":"80"},{"ruleId":"78","replacedBy":"81"},"no-native-reassign",["82"],"no-negated-in-lhs",["83"],["82"],["83"],"no-global-assign","no-unsafe-negation"]
+[{"C:\\Users\\magalu\\Desktop\\ConnectPets\\Front-End\\src\\components\\Pets\\index.js":"1","C:\\Users\\magalu\\Desktop\\ConnectPets\\Front-End\\src\\Pages\\Donate\\styles.js":"2","C:\\Users\\magalu\\Desktop\\ConnectPets\\Front-End\\src\\Pages\\Donate\\index.js":"3"},{"size":1148,"mtime":1617246144420,"results":"4","hashOfConfig":"5"},{"size":1471,"mtime":1617246730954,"results":"6","hashOfConfig":"5"},{"size":3285,"mtime":1617246795490,"results":"7","hashOfConfig":"5"},{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1037t6n",{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\magalu\\Desktop\\ConnectPets\\Front-End\\src\\components\\Pets\\index.js",[],"C:\\Users\\magalu\\Desktop\\ConnectPets\\Front-End\\src\\Pages\\Donate\\styles.js",[],"C:\\Users\\magalu\\Desktop\\ConnectPets\\Front-End\\src\\Pages\\Donate\\index.js",[]]
